requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.18.0
huggingface-hub>=0.15.0
datasets>=2.12.0
soundfile>=0.12.1
moviepy==1.0.3
numpy>=1.24.0
pillow>=9.5.0
accelerate>=0.20.0
safetensors>=0.3.1
torchvision>=0.15.0
torchaudio>=2.0.0
sentencepiece>=0.1.99
scipy>=1.10.0
dotenv
huggingface_hub[hf_xet]
optimum[onnxruntime]
streamlit
grpcio
grpcio-tools
protobuf
pydub
 